Rename database layers to include _db prefix
